fix: Add numeric type casting to production calculations for decimal precision
- dbt_dental_models/models/marts/mart_appointment_summary.sql: Add
  ::numeric casting to production_efficiency and
  avg_production_per_completed_appointment calculations. Prevents integer
  division truncation and ensures decimal precision for financial metrics
  (percentages and dollar amounts). Ensures accurate financial reporting
  with proper decimal places (90.00% not 90, $300.00 not $300).

+++ dbt_dental_models/models/marts/mart_appointment_summary.sql
@@ -261,8 +261,8 @@ final as (
         -- Financial Performance
         ae.total_scheduled_production,
         ae.completed_production,
-        round(ae.completed_production::numeric / nullif(ae.total_scheduled_production, 0) * 100, 2) as production_efficiency,
-        round(ae.completed_production / nullif(ae.completed_appointments, 0), 2) as avg_production_per_completed_appointment,
+        round((ae.completed_production::numeric / nullif(ae.total_scheduled_production, 0) * 100)::numeric, 2) as production_efficiency,
+        round((ae.completed_production::numeric / nullif(ae.completed_appointments, 0))::numeric, 2) as avg_production_per_completed_appointment,
         
         -- Performance Categories (moved to enhanced CTE)
         ae.completion_performance,
